+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,8 @@
 
 ## 9.11
 
+- **NEW**: Emoji: Update to just new CDN.
+- **NEW**: Emoji: Update to use Twemoji 14.1.2.
 - **NEW**: Snippets: Ignore nested snippet section syntax when including a section.
 
 ## 9.10